requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==6.7
Flask==1.0
Flask-OAuth==0.12
Flask-SSLify==0.1.5
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.1.1
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==16.8
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyodbc==4.0.30
pyparsing==2.2.0
python-http-client==2.2.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==0.15.3
